requirements.txt
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
pandas==3.0.1
python-dotenv==1.2.2
groq==1.1.0
streamlit>=1.40.0
